client/requirements.txt
altair==5.0.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
blinker==1.6.2
bokeh==3.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.41.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.3
linkify-it-py==2.0.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
networkx==3.1
numpy==1.25.0
packaging==23.1
pandas==2.0.3
panel==1.2.0
param==1.13.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.1
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==4.23.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyvista==0.40.1
pyviz-comms==2.3.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
rich==13.4.2
scikit-image==0.21.0
scipy==1.11.1
scooby==0.7.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
stpyvista==0.0.8
streamlit==1.24.0
tenacity==8.2.2
tifffile==2023.7.10
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trimesh==3.22.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
urllib3==2.0.3
validators==0.20.0
vtk==9.2.6
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
xyzservices==2023.7.0
zipp==3.15.0
